% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_location_names.R
\docType{data}
\name{key_location_names}
\alias{key_location_names}
\title{Master key for location names}
\format{
\subsection{\code{key_location_names}}{

A data frame with 2 columns:
\describe{
\item{location_name_unique}{Original location name.}
\item{location_name_concise}{The associated concise location name with typos and duplicates removed.}
}
}
}
\usage{
key_location_names
}
\description{
These data serve as a key to cross reference the unique (and sometimes duplicated) location names of sampling sites
and a set of 'concise' names that clean up characters and allow combination of duplicated locations.
}
\keyword{datasets}
